Enhance SATSolver with branching, logging, and progress
Added branching logic, progress tracking, and detailed logging to SATSolver. Introduced stream deletion and addition counters, improved stream management, and integrated Z3Solver for fast checks. Updated SATStream with copy constructor and fixed Z3Solver timeout type. Program now outputs solver logs and solution.

+++ SATStreams/Z3Solver.cs
@@ -17,7 +17,7 @@ namespace SATStreams
         {
             context = new Context();
             solver = context.MkSolver();
-            solver.Set("timeout", timeOut);
+            solver.Set("timeout", (uint)timeOut);
 
             var variables = Utils.GetVariables(cnf);
             varsDict = variables.ToDictionary(
